keep the convention

+++ src/applications/gqrx/remote_control.h
@@ -63,7 +63,7 @@ typedef struct {
     qint64 begin;             // beginning of the range
     qint64 step;              // step size in Hz
     size_t size;              // number of steps
-    double snr[MAP_ELEMENTS]; // signal to noise ratio per step
+    float snr[MAP_ELEMENTS]; // signal to noise ratio per step
     char   end;                 // end of the struct
 } snr_map_t;
